feat: change head only to head and src file separation
