+++ common/util.hpp
@@ -4,6 +4,7 @@
 #include <vector>
 #include <unordered_map>
 #include <fstream>
+#include <memory>
 #include <boost/algorithm/string.hpp>
 
 namespace hurrydocgo
