+++ controller/Makefile
@@ -1,6 +1,6 @@
 CC=g++
 FLAG=-std=c++17 -I ../third_party/cppjieba/include -lpthread -ljsoncpp -lboost_system
-hurrydocgo: controller.cpp ../index/index.cpp ../searcher/searcher.cpp
+hurrydocgo: controller.cpp ../index/index.cpp ../searcher/searcher.cpp ../preprocessor/preprocessor.cpp
 	$(CC) -o $@ $^ $(FLAG)
 .PHONY:clean
 clean:
